backend/requirements.txt
fastapi
uvicorn
django
orjson
zope.interface
venusian
wired
python-dotenv
python-jose